apps/email-update/pyproject.toml
[project]
name = "email-update"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.17.0",
    "services",
]

[tool.uv.sources]
services = { workspace = true }
